+++ packages/react-define-as-custom-element/src/defineAsCustomElementWithPortal.tsx
@@ -6,7 +6,10 @@ import createReactCustomElement from './private/createReactCustomElement.ts';
 import signalingState from './signalingState.ts';
 import { type AttributeAsProps, type AttributesMap, type DefineAsCustomElementInit } from './types.ts';
 
-type InstanceMapEntry<T extends object> = Readonly<[HTMLElement | ShadowRoot, Readonly<T>]>;
+type InstanceMapEntry<T extends object> = Readonly<
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  [HTMLElement | ShadowRoot, Readonly<T>, (fn: (name: string, ...args: any[]) => any) => void]
+>;
 type InstanceMap<T extends string> = ReadonlyMap<string, InstanceMapEntry<AttributeAsProps<T>>>;
 
 export default function defineAsCustomElement<T extends string>(
@@ -28,9 +31,12 @@ export default function defineAsCustomElement<T extends string>(
       constructor() {
         super(
           attributesMap,
+          init?.methodNames,
           init?.shadowRoot,
-          (props, element) =>
-            patchState(map => Object.freeze(new Map(map).set(this.#key, Object.freeze([element, props])))),
+          (props, element, setMethodCallback) =>
+            patchState(map =>
+              Object.freeze(new Map(map).set(this.#key, Object.freeze([element, props, setMethodCallback])))
+            ),
           () =>
             patchState(map => {
               const nextMap = new Map(map);
@@ -82,9 +88,14 @@ export default function defineAsCustomElement<T extends string>(
         {Array.from(
           instances
             .entries()
-            .map(([key, [element, props]]) =>
+            .map(([key, [element, props, setMethodCallback]]) =>
               createPortal(
-                <CustomElementProvider componentType={componentType} customElement={element} props={props} />,
+                <CustomElementProvider
+                  componentType={componentType}
+                  customElement={element}
+                  props={props}
+                  setMethodCallback={setMethodCallback}
+                />,
                 element,
                 key
               )