+++ packages/react-define-as-custom-element/src/index.ts
@@ -1,12 +1,14 @@
 import defineAsCustomElement from './defineAsCustomElement.tsx';
 import defineAsCustomElementWithPortal from './defineAsCustomElementWithPortal.tsx';
 import useCustomElement from './hooks/useCustomElement.ts';
+import useMethodCallback from './hooks/useMethodCallback.ts';
 import { type AttributeAsProps, type AttributesMap, type DefineAsCustomElementInit } from './types.ts';
 
 export {
   defineAsCustomElement,
   defineAsCustomElementWithPortal,
   useCustomElement,
+  useMethodCallback,
   type AttributeAsProps,
   type AttributesMap,
   type DefineAsCustomElementInit