+++ packages/react-define-as-custom-element/src/private/createReactCustomElement.ts
@@ -1,6 +1,11 @@
 import { type AttributeAsProps, type AttributesMap } from '../types.ts';
 
-type MountCallback<P extends object> = (props: P, element: HTMLElement | ShadowRoot) => void;
+type MountCallback<P extends object> = (
+  props: P,
+  element: HTMLElement | ShadowRoot,
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  setMethodCallback: (fn: (name: string, ...args: any[]) => any) => void
+) => void;
 type UnmountCallback = (element: HTMLElement | ShadowRoot) => void;
 
 export default function createReactCustomElement<T extends string>(
@@ -11,6 +16,7 @@ export default function createReactCustomElement<T extends string>(
   return class ReactCustomElement extends (customElementConstructor || HTMLElement) {
     constructor(
       attributesMap: AttributesMap<T>,
+      methodNames: string[] | undefined,
       shadowRootInit: ShadowRootInit | undefined,
       mountCallback: MountCallback<AttributeAsProps<T>>,
       unmountCallback: UnmountCallback
@@ -23,14 +29,30 @@ export default function createReactCustomElement<T extends string>(
 
       this.#element = shadowRootInit ? this.attachShadow(shadowRootInit) : this;
 
+      for (const methodName of methodNames || []) {
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        (this as any)[methodName] = (...args: any[]) => {
+          const callback = this.#methodCallback;
+
+          if (!callback) {
+            throw new Error(`useMethodCallback('${methodName}') must be called before this method can be called.`);
+          }
+
+          return callback(methodName, ...args);
+        };
+      }
+
       registry.register(new WeakRef(this), this[Symbol.dispose].bind(this));
     }
 
     #attributesMap: AttributesMap<T>;
     #connected: boolean = false;
     #element: ReactCustomElement | ShadowRoot;
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    #methodCallback: ((name: string, ...args: any) => any) | undefined;
     #mountCallback: MountCallback<AttributeAsProps<T>>;
     #propsMap: Map<T, string | undefined> = new Map();
+    #setMethodCallbackBound = this.#setMethodCallback.bind(this);
     #unmountCallback: UnmountCallback | undefined;
     #version: number = 0;
 
@@ -51,10 +73,15 @@ export default function createReactCustomElement<T extends string>(
           if (version === this.#version) {
             const element = this.#element;
 
-            element && this.#mountCallback(this.#getProps(), element);
+            element && this.#mountCallback(this.#getProps(), element, this.#setMethodCallbackBound);
           }
         });
       }
+    }
+
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    #setMethodCallback(fn: ((name: string, ...args: any) => any) | undefined) {
+      this.#methodCallback = fn;
     }
 
     attributeChangedCallback(name: string, _oldValue: string | undefined, newValue: string | undefined) {