+++ packages/react-define-as-custom-element/src/types.ts
@@ -11,6 +11,7 @@ export type DefineAsCustomElementInit = Readonly<{
         extends: string;
       }
     | undefined;
+  methodNames?: string[] | undefined;
   shadowRoot?: ShadowRootInit | undefined;
 }>;