+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Attribute changes will be batched via [`queueMicrotask`](https://developer.mozilla.org/en-US/docs/Web/API/Window/queueMicrotask), by [@compulim](https://github.com/compulim), in PR [#8](https://github.com/compulim/react-define-as-custom-element/pull/8)
+- Support custom methods by registering via `useMethodCallback()`, by [@compulim](https://github.com/compulim), in PR [#9](https://github.com/compulim/react-define-as-custom-element/pull/9)
 
 ### Removed